test(sdk/kanban): consolidate board concurrency + restore tests into board_test.go
Restore the test-file consolidation that was accidentally reverted while
amending the L5 commit. The concurrency and restore suites belong with
the rest of the Board tests; keeping them in dedicated files added an
extra layer of indirection without organisational benefit.

No production-code changes.

Made-with: Cursor
