+++ sdk/kanban/board_test.go
@@ -2,6 +2,8 @@ package kanban
 
 import (
 	"context"
+	"fmt"
+	"sync"
 	"testing"
 	"time"
 
@@ -939,6 +941,345 @@ func TestBoard_Bus_FanOutsToMultipleSubscribers(t *testing.T) {
 			}
 		case <-time.After(time.Second):
 			t.Errorf("sub[%d]: timed out", i)
+		}
+	}
+}
+
+// ---------------------------------------------------------------------------
+// RestoreBoard — rebuild a board from persisted KanbanCardModel rows.
+//
+// Bug 5 (issue #28): historical CreatedAt/UpdatedAt must survive restore so
+// timeline/SLA metrics keep reporting accurate elapsed times.
+// ---------------------------------------------------------------------------
+
+func TestRestoreBoard_AllStates(t *testing.T) {
+	t.Parallel()
+	now := time.Now()
+	cards := []*KanbanCardModel{
+		{
+			ID: "c1", RuntimeID: "s1", Type: "task", Status: "pending",
+			Producer: "copilot", Consumer: "*",
+			Query: "q1", TargetAgentID: "agent-1",
+			CreatedAt: now, UpdatedAt: now,
+		},
+		{
+			ID: "c2", RuntimeID: "s1", Type: "task", Status: "claimed",
+			Producer: "copilot", Consumer: "agent-2",
+			Query: "q2", TargetAgentID: "agent-2",
+			CreatedAt: now, UpdatedAt: now,
+		},
+		{
+			ID: "c3", RuntimeID: "s1", Type: "task", Status: "done",
+			Producer: "copilot", Consumer: "agent-3",
+			Query: "q3", TargetAgentID: "agent-3", Output: "result", RunID: "run-1",
+			CreatedAt: now, UpdatedAt: now,
+		},
+		{
+			ID: "c4", RuntimeID: "s1", Type: "task", Status: "failed",
+			Producer: "copilot", Consumer: "agent-4",
+			Query: "q4", TargetAgentID: "agent-4", Error: "timeout",
+			CreatedAt: now, UpdatedAt: now,
+		},
+	}
+
+	tb := RestoreBoard("s1", cards)
+	t.Cleanup(tb.Close)
+
+	if tb.ScopeID() != "s1" {
+		t.Fatalf("ScopeID = %q, want s1", tb.ScopeID())
+	}
+
+	got := make(map[string]string)
+	for _, c := range tb.Cards() {
+		got[c.ID] = c.Status
+	}
+	want := map[string]string{
+		"c1": "pending", "c2": "claimed", "c3": "done", "c4": "failed",
+	}
+	if len(got) != len(want) {
+		t.Fatalf("Cards()=%d, want %d", len(got), len(want))
+	}
+	for id, status := range want {
+		if got[id] != status {
+			t.Errorf("card %s: status=%q, want %q", id, got[id], status)
+		}
+	}
+
+	// Done card retains output / run_id during restore.
+	for _, c := range tb.Cards() {
+		if c.ID != "c3" {
+			continue
+		}
+		if c.RunID != "run-1" || c.Output != "result" {
+			t.Errorf("c3 lost done-state fields: %+v", c)
+		}
+	}
+}
+
+func TestRestoreBoard_Empty(t *testing.T) {
+	t.Parallel()
+	tb := RestoreBoard("s-empty", nil)
+	t.Cleanup(tb.Close)
+
+	if tb.ScopeID() != "s-empty" {
+		t.Fatalf("ScopeID = %q, want s-empty", tb.ScopeID())
+	}
+	if got := len(tb.Cards()); got != 0 {
+		t.Fatalf("Cards()=%d, want 0", got)
+	}
+}
+
+// Bug 5: historical timestamps must not be replaced by time.Now() during
+// the synthetic Claim/Done/Fail calls RestoreBoard makes internally.
+func TestRestoreBoard_PreservesTimestamps(t *testing.T) {
+	t.Parallel()
+	created := time.Date(2024, 1, 2, 10, 0, 0, 0, time.UTC)
+	updated := time.Date(2024, 1, 2, 10, 5, 30, 0, time.UTC)
+
+	tb := RestoreBoard("s1", []*KanbanCardModel{{
+		ID: "c-old", RuntimeID: "s1", Type: "task", Status: "done",
+		Producer: "copilot", Consumer: "agent-1",
+		Query: "q", TargetAgentID: "agent-1", Output: "ok",
+		CreatedAt: created, UpdatedAt: updated,
+	}})
+	t.Cleanup(tb.Close)
+
+	got, ok := tb.GetCardByID("c-old")
+	if !ok {
+		t.Fatal("restored card not found")
+	}
+	if !got.CreatedAt.Equal(created) {
+		t.Errorf("CreatedAt overwritten: got %v, want %v", got.CreatedAt, created)
+	}
+	if !got.UpdatedAt.Equal(updated) {
+		t.Errorf("UpdatedAt overwritten: got %v, want %v", got.UpdatedAt, updated)
+	}
+
+	// Derived fields (e.g. ElapsedMs) must reflect the restored stamps.
+	infos := tb.Cards()
+	if len(infos) != 1 {
+		t.Fatalf("Cards()=%d, want 1", len(infos))
+	}
+	wantElapsed := updated.Sub(created).Milliseconds()
+	if infos[0].ElapsedMs != wantElapsed {
+		t.Errorf("ElapsedMs = %d, want %d", infos[0].ElapsedMs, wantElapsed)
+	}
+}
+
+// WithTimestamps is the public hook RestoreBoard now uses internally.
+// Verify it overrides Produce's default time.Now() but does not break the
+// default path when not supplied.
+func TestBoard_WithTimestamps(t *testing.T) {
+	t.Parallel()
+	b := newBoard(t)
+
+	created := time.Date(2023, 6, 1, 12, 0, 0, 0, time.UTC)
+	updated := time.Date(2023, 6, 1, 12, 30, 0, 0, time.UTC)
+
+	t.Run("override", func(t *testing.T) {
+		card := b.Produce("task", "p", nil, WithTimestamps(created, updated))
+		if !card.CreatedAt.Equal(created) {
+			t.Errorf("CreatedAt = %v, want %v", card.CreatedAt, created)
+		}
+		if !card.UpdatedAt.Equal(updated) {
+			t.Errorf("UpdatedAt = %v, want %v", card.UpdatedAt, updated)
+		}
+	})
+
+	t.Run("default_still_now", func(t *testing.T) {
+		card := b.Produce("task", "p", nil)
+		if card.CreatedAt.IsZero() {
+			t.Fatal("default Produce must stamp CreatedAt")
+		}
+		if time.Since(card.CreatedAt) > time.Second {
+			t.Fatalf("default CreatedAt should be ~now, got %v ago", time.Since(card.CreatedAt))
+		}
+	})
+}
+
+// ---------------------------------------------------------------------------
+// Board concurrency — race / deadlock guards.
+//
+// Always run with -race. None of these tests assert ordering; they assert
+// invariants that must hold under any interleaving (no panics, all events
+// delivered, indexes stay consistent).
+// ---------------------------------------------------------------------------
+
+func TestBoard_Concurrent_ProduceAndQuery(t *testing.T) {
+	t.Parallel()
+	b := newBoard(t)
+	const n = 50
+
+	var wg sync.WaitGroup
+	for i := 0; i < n; i++ {
+		wg.Add(2)
+		go func() {
+			defer wg.Done()
+			b.Produce("task", "p", nil)
+		}()
+		go func() {
+			defer wg.Done()
+			_ = b.Query(CardFilter{Type: "task"})
+		}()
+	}
+	wg.Wait()
+
+	if got := b.Len(); got != n {
+		t.Fatalf("Len()=%d, want %d", got, n)
+	}
+}
+
+func TestBoard_Concurrent_FullLifecycle(t *testing.T) {
+	t.Parallel()
+	b := newBoard(t)
+	const n = 100
+
+	cardIDs := make([]string, n)
+	var wg sync.WaitGroup
+
+	for i := 0; i < n; i++ {
+		wg.Add(1)
+		go func(idx int) {
+			defer wg.Done()
+			c := b.Produce("task", fmt.Sprintf("p-%d", idx), map[string]any{"idx": idx})
+			cardIDs[idx] = c.ID
+		}(i)
+	}
+	wg.Wait()
+	if got := b.Len(); got != n {
+		t.Fatalf("after Produce: Len()=%d, want %d", got, n)
+	}
+
+	for i := 0; i < n; i++ {
+		wg.Add(1)
+		go func(idx int) {
+			defer wg.Done()
+			if !b.Claim(cardIDs[idx], fmt.Sprintf("c-%d", idx)) {
+				t.Errorf("claim %d failed", idx)
+			}
+		}(i)
+	}
+	wg.Wait()
+
+	for i := 0; i < n; i++ {
+		wg.Add(1)
+		go func(idx int) {
+			defer wg.Done()
+			b.Done(cardIDs[idx], fmt.Sprintf("result-%d", idx))
+		}(i)
+	}
+	wg.Wait()
+
+	if got := len(b.Query(CardFilter{Status: CardDone})); got != n {
+		t.Fatalf("done cards = %d, want %d", got, n)
+	}
+}
+
+func TestBoard_Concurrent_WatchAndProduce(t *testing.T) {
+	t.Parallel()
+	b := newBoard(t)
+	ctx, cancel := context.WithCancel(context.Background())
+	t.Cleanup(cancel)
+
+	const watchers = 5
+	const produces = 20
+
+	channels := make([]<-chan *Card, watchers)
+	for i := 0; i < watchers; i++ {
+		channels[i] = b.WatchFiltered(ctx, CardFilter{})
+	}
+
+	var wg sync.WaitGroup
+	for i := 0; i < produces; i++ {
+		wg.Add(1)
+		go func(idx int) {
+			defer wg.Done()
+			b.Produce("task", "p", map[string]any{"i": idx})
+		}(i)
+	}
+	wg.Wait()
+
+	for wi, ch := range channels {
+		count := 0
+		timeout := time.After(time.Second)
+	drain:
+		for {
+			select {
+			case <-ch:
+				count++
+				if count == produces {
+					break drain
+				}
+			case <-timeout:
+				break drain
+			}
+		}
+		if count != produces {
+			t.Fatalf("watcher %d: got %d events, want %d", wi, count, produces)
+		}
+	}
+}
+
+func TestBoard_Concurrent_BusSubscribe(t *testing.T) {
+	t.Parallel()
+	b := newBoard(t)
+
+	ctx := context.Background()
+	const subs = 20
+
+	var wg sync.WaitGroup
+	errs := make(chan error, subs)
+	for i := 0; i < subs; i++ {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			if _, err := b.Bus().Subscribe(ctx, event.Pattern(">")); err != nil {
+				errs <- err
+			}
+		}()
+	}
+	wg.Wait()
+	close(errs)
+	for err := range errs {
+		t.Fatalf("concurrent Subscribe failed: %v", err)
+	}
+}
+
+// CountByStatus must remain consistent with Query() under concurrent
+// state transitions — guards the cardIndex / statusIndex pair.
+func TestBoard_Concurrent_IndexConsistency(t *testing.T) {
+	t.Parallel()
+	b := newBoard(t)
+	const n = 50
+
+	ids := make([]string, n)
+	for i := 0; i < n; i++ {
+		ids[i] = b.Produce("task", "p", nil).ID
+	}
+
+	if got := b.CountByStatus(CardPending, ""); got != n {
+		t.Fatalf("initial pending=%d, want %d", got, n)
+	}
+
+	done := make(chan struct{})
+	for i := 0; i < n; i++ {
+		go func(id string) {
+			b.Claim(id, "a")
+			b.Done(id, "r")
+			done <- struct{}{}
+		}(ids[i])
+	}
+	for i := 0; i < n; i++ {
+		<-done
+	}
+
+	for status, want := range map[CardStatus]int{
+		CardDone:    n,
+		CardPending: 0,
+		CardClaimed: 0,
+	} {
+		if got := b.CountByStatus(status, ""); got != want {
+			t.Errorf("CountByStatus(%s)=%d, want %d", status, got, want)
 		}
 	}
 }
